requirements.txt
anytree==2.12.1
Cheetah3==3.2.6.post1
lxml==5.1.0
